+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## General
 
+* `ggpredict()` now supports models of class `rqs` from package *quantreg*.
+
 * Fixed issues to be compatible with forthcoming update of *emmeans*.
 
 # ggeffects 1.3.3